adhd-bot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-telegram-bot==21.9
google-cloud-firestore==2.19.0
google-cloud-tasks==2.16.5
google-cloud-secret-manager==2.21.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0
vertexai==1.74.0
stripe==11.4.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-mock==3.14.0
anyio==4.8.0
